+++ script/build.js
@@ -1,19 +1,15 @@
 #!/usr/bin/env node
 
 import sh from './sh.js';
-// import childProcess from 'node:child_process';
-// import fs from 'node:fs';
 
 // const tgtDir = `./out/src`;
 // const cwd = process.cwd();
 
 // Remove the build directory
 sh.rm('./out/src');
-// fs.rmSync(tgtDir, {force: true, recursive: true});
 
 // Make new build directory
 sh.mkdir('./out/src');
-// fs.mkdirSync(tgtDir, {recursive: true});
 
 // Compile the data through babel and put the result into the build directory
 sh.exec('./node_modules/.bin/babel --out-dir out/src src');
@@ -21,4 +17,3 @@ sh.exec('./node_modules/.bin/babel --out-dir out/src src');
 
 // Make the Client-Runner processable
 sh.chmod('./out/src/ESDocCLI.js', '755');
-// fs.chmodSync(`${tgtDir}/ESDocCLI.js`, 755);