+++ script/test-ci.js
@@ -2,6 +2,8 @@
 
 import sh from './sh.js';
 
-sh.exec('./script/eslint.js');
-sh.exec('./script/test.js --coverage');
-sh.exec('./node_modules/.bin/codecov');
+const cwd = process.cwd();
+
+sh.exec(`${cwd}/script/eslint.js`);
+sh.exec(`${cwd}/script/test.js --coverage`);
+// sh.exec(`${cwd}/node_modules/.bin/codecov`);