+++ src/Parser/ESParser.js
@@ -1,7 +1,5 @@
-// Object.defineProperty(exports, "__esModule", {value: true});
-
-import babylon from 'babylon';
-import fsExtra from 'fs-extra';
+import babelParser from '@babel/parser';
+import fs from 'node:fs';
 
 import Plugin from '../Plugin/Plugin.js';
 
@@ -26,12 +24,6 @@ import Plugin from '../Plugin/Plugin.js';
 //   return result;
 // }
 
-// function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : {def: obj}; }
-
-// const babylon = _interopRequireWildcard(babylon);
-// const fsExtra2 = _interopRequireDefault(fsExtra);
-// const _Plugin2 = _interopRequireDefault(_Plugin);
-
 /**
  * ECMAScript Parser class.
  *
@@ -45,7 +37,7 @@ export default class ESParser {
    * @returns {AST} AST of source code.
    */
   static parse(filePath) {
-    let code = fsExtra.readFileSync(filePath, {encode: `utf8`}).toString();
+    let code = fs.readFileSync(filePath, {encode: `utf8`}).toString();
     code = Plugin.onHandleCode(code, filePath);
     if (code.charAt(0) === `#`) code = code.replace(/^#!/u, `//`);
 
@@ -53,7 +45,7 @@ export default class ESParser {
       plugins:    [],
       sourceType: `module`
     };
-    let parser = code => babylon.parse(code, parserOption);
+    let parser = code => babelParser.parse(code, parserOption);
 
     ({parser, parserOption} = Plugin.onHandleCodeParser(parser, parserOption, filePath, code));
 
@@ -64,4 +56,3 @@ export default class ESParser {
     return ast;
   }
 }
-// exports.default = ESParser;